% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wobble.R
\name{as_wobble}
\alias{as_wobble}
\title{As wobble}
\usage{
as_wobble(.x, .n)
}
\arguments{
\item{.x}{Input wibble}

\item{.n}{Node on which to set observation level}
}
\value{
A wobble with attributes and .node_id as initial variables
}
\description{
Convert wibbles to wobbles data frame with a set observation level
}
