% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{coerce_tbls}
\alias{coerce_tbls}
\title{Coerce list into list of data frames}
\usage{
coerce_tbls(x)
}
\arguments{
\item{x}{Input list}
}
\description{
Combines same-length elements of input list into data frames
}
