% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_nodes.R
\name{h_node}
\alias{h_node}
\title{h_node}
\usage{
h_node(.x, .n)
}
\arguments{
\item{.x}{Input nodes object}

\item{.n}{CSS used to identify the node(s)}
}
\value{
A node object
}
\description{
Identify and return node object via CSS
}
