% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_nodes.R
\name{h_nodes}
\alias{h_nodes}
\title{h_nodes}
\usage{
h_nodes(.x, .n)
}
\arguments{
\item{.x}{Input nodes object}

\item{.n}{CSS used to identify the node(s)}
}
\value{
Nodes or list of nodes object
}
\description{
Identify and return node objects via CSS
}
