% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_text.R
\name{h_text}
\alias{h_text}
\title{h_text}
\usage{
h_text(.x, trim = TRUE)
}
\arguments{
\item{.x}{Input nodes object}

\item{...}{Passed to xml2::xml_text()}
}
\value{
Character vector or list of character vectors
}
\description{
Identify and return node object via CSS
}
