% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wutate.R
\name{node_chr}
\alias{node_chr}
\title{Character node variable}
\usage{
node_chr(.n)
}
\arguments{
\item{.n}{A string representing the node from which content should be
extracted.}
}
\value{
A character vector
}
\description{
Returns a character vector from the first node of each element of input
}
