% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wutate.R
\name{nodes_chr}
\alias{nodes_chr}
\title{Character nodes variable}
\usage{
nodes_chr(.n)
}
\arguments{
\item{.n}{A string representing the nodes from which content should be
extracted for each input element.}
}
\value{
A character vector
}
\description{
Returns a character vector from a pasted together-string of all nodes for
each element of input
}
