% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{Parse JSON values from object}
\usage{
parse_json(x)
}
\arguments{
\item{x}{Input object}
}
\value{
Parsed JSON (if detected)
}
\description{
Detects and parses JSON text into R objects
}
\examples{
\dontrun{
## parse JSON text found on espn's homepage
e <- parse_json("http:/www.espn.com/")

## view parsed object
str(e, 3)

## parse JSON from html string
parse_json('<html>{\\"mpg\\":21,\\"cyl\\":6,\\"disp\\":160,\\"lgl\\":false}</html>')
}
}
