% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wibble.R
\name{read_wibble}
\alias{read_wibble}
\title{Read wibble}
\usage{
read_wibble(x)
}
\arguments{
\item{x}{URL, file path, or character vector of source code}
}
\description{
Read content as wibble
}
\details{
This is a wrapper around xml2's read function.
}
