% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{tbl_and_bind}
\alias{tbl_and_bind}
\title{Convert to tbl and then bind rows}
\usage{
tbl_and_bind(x)
}
\arguments{
\item{x}{Input list}
}
\value{
A list of data frames and lists
}
\description{
Convert lists to data frames then bind rows
}
