% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wabble.R
\name{wabble}
\alias{wabble}
\title{Wabble}
\usage{
wabble(url, query = NULL, path = "")
}
\arguments{
\item{url}{Base URL}

\item{query}{Named list of query parameters}

\item{path}{Endpoint}
}
\value{
A wbl_url
}
\description{
API/URL builder
}
\examples{

## query params as list
params <- list(sort = 'desc', page = "4")

## build URL
wabble("https://mikewk.com", params, path = "search.json")

}
