% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpage.R
\docType{data}
\name{Webpage}
\alias{Webpage}
\title{Webpage object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Webpage
}
\description{
An R6 class to represent source code/data associated with a URL (website).
}
\section{Functions}{

Some text here

\describe{
\item{one}{description}
}
}

\keyword{datasets}
