% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wibble-package.R
\docType{package}
\name{wibble-package}
\alias{wibble}
\alias{wibble-package}
\title{wibble: Web Data Frames}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Wibble provides a simple and streamlined interface for retrieving,
    parsing, and wrangling unstructured web data with a particular emphasis on
    supplying tools for converting web scraped-information into tidy data frames.
}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
