% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wibble.R
\name{wibble}
\alias{wibble}
\title{Wibble}
\usage{
wibble(.x)
}
\arguments{
\item{.x}{Input object}
}
\value{
A wbl data frame
}
\description{
A web data frame
}
\examples{

## return wbl_df
wibble(list(a = 1:5, b = TRUE, c = NULL))

}
