% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobble.R
\name{wobble}
\alias{wobble}
\title{Wobble}
\usage{
wobble(.x, .n)
}
\arguments{
\item{.x}{A wbl}

\item{.n}{Node on which to set observation level}
}
\value{
A wbl_df
}
\description{
Set observation (row)-level for wibble
}
