% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wutate.R
\name{wutate}
\alias{wutate}
\title{Mutate web data}
\usage{
wutate(.data, ...)
}
\arguments{
\item{.data}{A wibble}

\item{...}{Expressions resolving as variables in .data}
}
\value{
A wibble
}
\description{
Transform and add variables to wibble
}
